Cleaned up codec code
